Develop

+++ src/ArduinoHardwareSerial.h
@@ -2,8 +2,9 @@
 // Created by Anton Matosov on 5/4/16.
 //
 
-#pragma once 
+#pragma once
 
+#include <Arduino.h>
 #include <HAL.h>
 
 class HardwareSerial;
@@ -14,7 +15,7 @@ public:
     ArduinoHardwareSerial(HardwareSerial& arduinoSerial, uint8_t rxdPin);
 
     void begin(const unsigned long baudRate, const uint8_t transferConfig = SERIAL_8N1) override;
-    unsigned int write(uint8_t byte) override;
+    size_t write(uint8_t byte) override;
     bool available() override;
     uint8_t peek() override;
     uint8_t read() override;
